backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.117.1",
    "google-api-core>=2.25.1",
    "gtts>=2.5.4",
    "jose>=1.0.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-google-genai>=4.0.0",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "psycopg2>=2.9.11",
    "pydantic[email]>=2.11.9",
    "pydantic-settings>=2.12.0",
    "pygame>=2.6.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.1",
    "speechrecognition>=3.14.3",
    "sqlalchemy>=2.0.43",
    "sse-starlette>=3.0.2",
    "tenacity>=9.1.2",
    "uvicorn>=0.37.0",
    "pymongo>=4.6.0",
    "langchain-community>=0.4.1",
    "unstructured[pdf]>=0.18.27",
    "chromadb>=1.1.0",
    "transformers>=4.57.5",
]
